requirements.txt
Flask==1.0.2
py-cpuinfo==4.0.0
gunicorn==19.9.0
